Refactor notification system, use notifications table
Instead of querying the notifications directly from the activities
table, we introduce a notifications table holds a record
for each individual notification. In order to query the
notifications for a single user, a simple select on the
notificitions table has to be done.

The design is much more flexible and simpler, however
DB writes increases heavily and should be defered to a background
process in the future. But since we implement an email solution
as well, we have to come up with a background worker stack anyway.